Add Swift API endpoint
