+++ cmd/swift.go
@@ -5,9 +5,12 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"net/url"
 	"os"
+	"strconv"
 	"strings"
 
+	"github.com/cozy/cozy-stack/client/request"
 	"github.com/cozy/cozy-stack/model/instance"
 	"github.com/cozy/cozy-stack/model/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/config/config"
@@ -38,66 +41,27 @@ var swiftCmdGroup = &cobra.Command{
 
 var lsLayoutsCmd = &cobra.Command{
 	Use:     "ls-layouts",
-	Short:   `Count layouts by types (v1, v2a, v2b)`,
+	Short:   `Count layouts by types (v1, v2a, v2b, v3)`,
 	Example: "$ cozy-stack swift ls-layouts",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		type layout struct {
-			Counter int      `json:"counter"`
-			Domains []string `json:"domains,omitempty"`
-		}
-		var layoutV1, layoutV2a, layoutV2b, layoutUnknown, layoutV3 layout
-
-		instances, err := instance.List()
+		c := newAdminClient()
+		values := url.Values{}
+		values.Add("show_domains", strconv.FormatBool(flagShowDomains))
+		res, err := c.Req(&request.Options{
+			Method:  "GET",
+			Path:    "/swift/list-layouts",
+			Queries: values,
+		})
 		if err != nil {
 			return err
 		}
-		for _, inst := range instances {
-			switch inst.SwiftLayout {
-			case 0:
-				layoutV1.Counter++
-				if flagShowDomains {
-					layoutV1.Domains = append(layoutV1.Domains, inst.Domain)
-				}
-			case 1:
-				switch inst.DBPrefix() {
-				case inst.Domain:
-					layoutV2a.Counter++
-					if flagShowDomains {
-						layoutV2a.Domains = append(layoutV2a.Domains, inst.Domain)
-					}
-				case inst.Prefix:
-					layoutV2b.Counter++
-					if flagShowDomains {
-						layoutV2b.Domains = append(layoutV2b.Domains, inst.Domain)
-					}
-				default:
-					layoutUnknown.Counter++
-					if flagShowDomains {
-						layoutUnknown.Domains = append(layoutUnknown.Domains, inst.Domain)
-					}
-				}
-			case 2:
-				layoutV3.Counter++
-				if flagShowDomains {
-					layoutV3.Domains = append(layoutV3.Domains, inst.Domain)
-				}
-			default:
-				layoutUnknown.Counter++
-				if flagShowDomains {
-					layoutUnknown.Domains = append(layoutUnknown.Domains, inst.Domain)
-				}
-			}
+		defer res.Body.Close()
+
+		var buf interface{}
+		if err := json.NewDecoder(res.Body).Decode(&buf); err != nil {
+			return err
 		}
-
-		output := make(map[string]interface{})
-		output["v1"] = layoutV1
-		output["v2a"] = layoutV2a
-		output["v2b"] = layoutV2b
-		output["unknown"] = layoutUnknown
-		output["v3"] = layoutV3
-		output["total"] = layoutV1.Counter + layoutV2a.Counter + layoutV2b.Counter + layoutUnknown.Counter + layoutV3.Counter
-
-		json, err := json.MarshalIndent(output, "", "  ")
+		json, err := json.MarshalIndent(buf, "", "  ")
 
 		if err != nil {
 			return err
